Cleaned up init file. Everything must be done in  genexpert.py.
